Refactor Button component.
